requirements.txt
asgiref==3.2.3
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0
django-cors-headers==3.2.0
django-environ==0.4.5
django-extensions==2.2.5
django-model-utils==4.0.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
firebase-admin==3.2.0
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.7.2
google-auth-httplib2==0.0.3
google-cloud-core==1.1.0
google-cloud-firestore==1.6.0
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
grpcio==1.25.0
gunicorn==20.0.4
httplib2==0.14.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
msgpack==0.6.2
openapi-codec==1.3.2
phonenumbers==8.11.1
pkg-resources==0.0.0
protobuf==3.11.1
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyJWT==1.7.1
pytz==2019.3
PyYAML==5.2
requests==2.22.0
rsa==4.0
services-communicator==0.1.2
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
